requirements.txt
openai
python-telegram-bot==12.7
